requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==4.2.27
django-filter==25.1
djangorestframework==3.16.1
gunicorn==23.0.0
importlib_metadata==8.7.1
Markdown==3.9
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0
zipp==3.23.0
